requirements.txt
dj-database-url==0.3.0
Django==1.8.1
django-ckeditor==4.4.8
django-datetime-widget==0.9.3
django-endless-pagination==2.0
django-filebrowser==3.5.7
django-grappelli==2.6.5
django-postgrespool==0.3.0
django-tinymce==1.5.3
Pillow==2.8.1
psycopg2==2.6
pytz==2015.4
SQLAlchemy==1.0.4
virtualenv==13.0.3
wheel==0.24.0
whitenoise==1.0.6
WTForms==2.0.2
